CI: Remove some runners with Clang on Ubuntu
LLVM Clang emits a different set of compiler warnings for potential
issues in the source code (compared to GCC). So, it makes sense to keep
building with it in CI.
There is probably little reason (when it comes to coverage) to build
with different configurations with that compiler though.

Keep only one configuration that builds with Clang on Ubuntu (to save
some CI minutes - and reduce the energy footprint).