CI (alpine): Use Ubuntu on ARM runners for armv7.
The armv7 architecture can run natively on the aarch64 CPUs of the
GitHub-hosted Ubuntu on ARM runners which makes it a lot faster than
emulation on an x86_64 CPU. That allows to also build GraphBLAS and
LAGraph for that architecture in an acceptable amount of time now.

With thanks to @bgilbert for the instructions how to use the hosted
Ubuntu on ARM runners with the setup-alpine action.