CHOLMOD: Resolve CUDA device symbols when linking static library.
If CUDA device symbols are not resolved when linking a static library
(with MSVC) dependent binaries need to be linked with the CUDA compiler
as the linker driver. That doesn't happen automatically with current
versions of CMake.

Instead, resolve CUDA device symbols when linking the static CHOLMOD
library. That means that no CUDA calls will be possible from dependent
binaries to CUDA functions in the static library.
As far as I can tell, no CUDA functions are part of the API of the
CHOLMOD library. So, doing that shouldn't be an issue.