Support building SuiteSparse_config without BLAS.
BLAS is a dependency of SuiteSparse_config for some helper functions that
can be used in other SuiteSparse libraries. If no SuiteSparse library
that depends on BLAS is build, these helper functions are not used, and
they are allowed to return anything in general without potential adverse
side effects.

Add a new configuration flag SUITESPARSE_REQUIRE_BLAS that can be used to
manually remove the requirement for a BLAS implementation when building
SuiteSparse_config.